+++ ext/crr/src/changes-vtab-write.c
@@ -7,7 +7,6 @@
 #include "consts.h"
 #include "crsqlite.h"
 #include "ext-data.h"
-#include "seen-peers.h"
 #include "tableinfo.h"
 #include "util.h"
 
@@ -283,8 +282,6 @@ int crsql_mergeInsert(sqlite3_vtab *pVTab, int argc, sqlite3_value **argv,
                                                  pTab->pExtData->tableInfosLen,
                                                  (const char *)insertTbl);
 
-  crsql_trackSeenPeer(pTab->pSeenPeers, insertSiteId, insertSiteIdLen,
-                      insertDbVrsn);
   if (tblInfo == 0) {
     *errmsg = sqlite3_mprintf(
         "crsql - could not find the schema information for table %s",