+++ ext/crr/src/ext-data.h
@@ -11,7 +11,6 @@ struct crsql_ExtData {
   // perma statement -- used to check db schema version
   sqlite3_stmt *pPragmaSchemaVersionStmt;
   sqlite3_stmt *pPragmaDataVersionStmt;
-  sqlite3_stmt *pTrackPeersStmt;
   int pragmaDataVersion;
 
   // this gets set at the start of each transaction on the first invocation