+++ ext/crr/src/tests.c
@@ -22,7 +22,6 @@ void crsqlChangesVtabTestSuite();
 void crsqlChangesVtabWriteTestSuite();
 void crsqlChangesVtabCommonTestSuite();
 void crsqlExtDataTestSuite();
-void crsqlSeenPeersTestSuite();
 void crsqlFractSuite();
 void crsqlIsCrrTestSuite();
 
@@ -40,7 +39,6 @@ int main(int argc, char *argv[]) {
   SUITE("vtabwrite") crsqlChangesVtabWriteTestSuite();
   SUITE("vtabcommon") crsqlChangesVtabCommonTestSuite();
   SUITE("extdata") crsqlExtDataTestSuite();
-  SUITE("seenpeers") crsqlSeenPeersTestSuite();
   // integration tests should come at the end given fixing unit tests will
   // likely fix integration tests
   SUITE("crsql") crsqlTestSuite();